Added shooting placeholder mechanic

+++ scripts/Player.js
@@ -32,38 +32,47 @@ Player.prototype.handleEvent = function(e) {
     // the key pressed
     var code = e.keyCode;
 
-    // if key pressed is not in our keymap, exit
-    if(!(code in keyMap)) { return; }
-    
-    // ???
-    // profit
-    var diff = ROT.DIRS[8][keyMap[code]];
-    var newX = this._x + diff[0];
-    var newY = this._y + diff[1];
-    
-    var currentKey = this._x+','+this._y;
-    var newKey = newX + "," + newY;
-    if (!(newKey in Game.map)) { return; }
-    var glyph = Game.map[newKey].getGlyph();
-    if (!glyph.isPassable()) { // wall or something, can't move here
-            return;
-    } else if (newKey in Game.entities) { // kill it!
-        //TODO: implement attacking
-        this._attack(Game.entities[newKey]);
-        Game.engine.unlock();
-    } else { // move there
-        var glyph = Game.map[currentKey].getGlyph();
-        Game.display.draw(this._x, this._y, glyph.getChar(), glyph.getForeground(), glyph.getBackground());
-        this._x = newX;
-        this._y = newY;
-        this.draw();
-        window.removeEventListener("keydown", this);
-        Game.engine.unlock();
+    if (code == 83) { 
+        this._shoot();
+    } else {
+        // if key pressed is not in our keymap, exit
+        if(!(code in keyMap)) { return; }
+        
+        // ???
+        // profit
+        var diff = ROT.DIRS[8][keyMap[code]];
+        var newX = this._x + diff[0];
+        var newY = this._y + diff[1];
+        
+        var currentKey = this._x+','+this._y;
+        var newKey = newX + "," + newY;
+        if (!(newKey in Game.map)) { return; }
+        var glyph = Game.map[newKey].getGlyph();
+        if (!glyph.isPassable()) { // wall or something, can't move here
+                return;
+        } else if (newKey in Game.entities) { // kill it!
+            //TODO: implement attacking
+            this._attack(Game.entities[newKey]);
+            Game.engine.unlock();
+        } else { // move there
+            var glyph = Game.map[currentKey].getGlyph();
+            Game.display.draw(this._x, this._y, glyph.getChar(), glyph.getForeground(), glyph.getBackground());
+            this._x = newX;
+            this._y = newY;
+            this.draw();
+            window.removeEventListener("keydown", this);
+            Game.engine.unlock();
+        }
     }
 };
 
 Player.prototype._attack = function(entity) {
     entity.destroy();
+}
+
+Player.prototype._shoot = function() {
+    /* TODO: Make shoot do more than make pew pew sounds */
+    console.log('pew pew');
 }
 
 Player.prototype.getX = function() { return this._x; }
